Fix validation of DRM modifier compatibility with image properties
Vulkan spec states that in case of image creation with list of
DRM modifiers (VkImageDrmFormatModifierListCreateInfoEXT), each
of them must be compatible with the parameters in
VkImageCreateInfo and its pNext chain.

This change adds a mandatory check for modifier compatibility with
the specified image properties by querying
vkGetPhysicalDeviceImageFormatProperties in a place where it was
missing prior to the actual image creation.

Components: Vulkan

Affects: dEQP-VK.drm_format_modifiers.create_list_modifiers.*

VK-GL-CTS Issue: 2611

Change-Id: Iab5deff6c8cbe6d2c510224322f7be91c7874acf