Add subpass self dependency test with a single attachment
This test replicates an issue seen in an app that behaves incorrectly
on a specific driver implementation. It uses a single attachment
referenced as both an input and an output. There is also alpha blending
enabled even though it has no effect in this particular test case.
This was needed to reproduce the original bug.

VK-GL-CTS Issue: 2597

New tests:

dEQP-VK.renderpass*.suballocation.subpass_dependencies.single_attachment.*

Components: Vulkan
Change-Id: Iaf79b3937dec4eebae1b6ad7ec035c1bdf05e811