Check for VK_KHR_portability_subset extension before checking it's features
Fix text 'not supported' logic, by checking if VK_KHR_portability_subset
extension is supported. If it's not, than it's features should not be checked
and the test should just continue.

This is a fix for a regression introduced in 5534dd8f

Components: Vulkan

VK-GL-CTS Issue: 2711

Affects: dEQP-VK.image.extended_usage_bit.*

Change-Id: Idbb0d6dd10bc21b7882f388757beceaa5af80ea6