Fix undefined behavior in FloatToHalf
The size of the exponent and mantissa in a GLhalf is smaller than
a GLfloat. This means that small numbers (that approach 0.0) will
become 0.0. This can result in doing a right-shift by 32 or more bits,
which is undefined behavior. It works with some older compilers/CPUs,
but not with newer clang.

Affects: KHR-GLES31.core.vertex_attrib_binding.basic-input-case*

Components: OpenGL ES

VK-GL-CTS issue: 2723

Change-Id: I0b1d309c6919576ef63d19f72bca37254ca11b8f