Stop dEQP-GL45 tests falling back to ES3.1
The ES3.1 context creation code accepted an apiType parameter but would
always fall back to ES3.1 if the apiType that was requested was not
supported. (This used to be used to create an ES3.2 context where
supported). This meant that the new dEQP-GL45.* tests would run on an ES
implementation using an ES3.1 context.

Make the behaviour of dEQP-GL45.* consistent with the KHR-GL* tests by
failing to create a context if the requested apiType is not supported.
The old fallback behaviour is moved to the dEQP-GLES31 test package,
which is where the strange fallback requirement comes from anyway.

Components: OpenGL
VK-GL-CTS issue: 2747
Affects: dEQP-GL45.*

Change-Id: Ie5ef92c708a28211a007589c0f4aa49ad7d745e1
(cherry picked from commit ff4410749c669b9b6e943a41f76d30b6db86d034)