Verify submit with wait on timeline semaphore with 0 value
We have test that verifies a vkWaitSemaphores() on a timeline
semaphore value of 0. Because a timeline semaphore always has a 0
value initially, the wait should return immediately.

We can also verify that a vkQueueSubmit on the same semaphore value is
not blocked on anything.

VK-GL-CTS Issue: 1827

Affects: dEQP-VK.synchronization.timeline_semaphore.device_host.misc.initial_value

Change-Id: Ifcfe053206e9f5f836aa25c7b6edd8fa7fdfb540