Shared memory layout tests
Vulkan leaves it to implementations to determine the layout of shared memory.
This test ensures that the fields in the shared memory won't end up
overlapping each other.

This commit also does a minor clean up for the framework core and adds
util functions for comparing types which are common for this and the SSBO tests.

VK-GL-CTS issue: 2372

New Tests:
dEQP-VK.memory_model.shared.*

Components: Vulkan

Change-Id: I4ac0a03f91a5ce699fe03f3c43943fb8eac88850