Fix precision issue in ShaderMatrixTest
According to the Spec, the compiler is free to promote computation
precisions, and such promotion depends on the specific shader
content or compiler.

When tests use the lowp or mediump precision in the shaders, It
can't confirm the final computation precision. Now the precision
of the shader inputs is highp, if the real computing precision in
GPU HW is lowp or mediump. It can easily get the comparison failure.
So it's better to adjust the precision of the inputs as the expected
precision in the specific tests.

Affects:

dEQP-GLES3.functional.shaders.matrix.*

Components: OpenGL

VK-GL-CTS issue: 2863

Change-Id: I30af699da12bf88596e7046922b2f1f00d5d3f06