Implement MinGW build support
This commit adds support for building with the MinGW toolchain.

This is useful as it allows Windows builds of the CTS to be
built on Linux or other platforms.
This can then either be used on Windows natively, or under the Wine
translation layer.

It can be used by either specifying a MinGW toolchain in CMake
or using the x86_64-w64-mingw32-cmake wrapper.

Fixes: #125

Signed-off-by: Joshua Ashton <joshua@froggi.es>