Add a new GLES2TextureSharingTest
This is a test for texture used in default fbo and user fbo
in same context. If this texture sampled in default fbo and
might be rendered in user fbo, the texture buffer might be
wrongly allocated which cause two rendering results different.

The test's behavior is like this:
1, draw left half rectangle, sample the GL texture
2, attach the GL texture as color attachment, read 1 pixel
3, draw right half rectangle, sample the GL texture
4, the first time drawing result as reference and compare with
the second time rendering result.

Affects:
dEQP-EGL.functional.sharing.gles2.texture.render_sample_mixed

Components: OpenGL

Change-Id: I11e0fb5745588d5e5440a22a540b7651976c934d