merge consecutive lines when stippling
When we draw the reference images, we split line segments into
single-pixel length line segment, and we consider the stipple pattern
for each segment individually. This leads to close to every covered
pixel being considered an edge pixel, because we draw rectangles that
are a single pixel long.

We should merge the consecutive line segments before rasterizing them,
so the filled portions are considered as a whole.

Without this, the line-stippling tests are close to pointless; very
close to every single covered pixel is allowed to be either filled or
not, leading to very very few pixels actually being tested.