Add missing checking of integer dot product features
The requested features for dot products were being set up by the tests
but they were never actually validated. (In practice this probably won't
be a problem because they are compulsory with the extension and also in
the new core version, but it makes the code quite confusing).

Components: Vulkan
Affects: dEQP-VK.spirv_assembly.*

Change-Id: I62558d27f9bb360e3912b90964739bb1356df4b8