Fix missing barrier in FSR pixel consistency tests
The PixelConsistency tests of Fragment Shading Rate did not have a
pipeline barrier between the calls to vkCmdClearColorImage for the
render targets and the subsequent renderpass. This caused a race between
the two.

This commit adds the missing pipeline barriers.
In addition, it converts the preImageBarrier from two calls to
vkCmdPipelineBarrier with a single barrier to a single call with two
barriers.

Components: Vulkan

Affects: dEQP-VK.fragment_shading_rate.pixel_consistency.*

VK-GL-CTS Issue: 3158

Change-Id: I14b5de4fb02593b0a384b0f36892543ee2c57222