Port deqp-vk to Fuchsia
"Fuchsia is a modern open source operating system that's simple, secure,
updatable, and performant." See fuchsia.dev for details.

This change only modifies the code so it can be build in a
Fuchsia-specific environment. Build rules depend on the exact version of
the Fuchsia SDK in use, and will be added later.

VK-GL-CTS Issue: 3849

Components: Framework
Change-Id: Ia869ec251ba8392c2a6712859e33c38bea88a099