Use standard type definitions on all platforms
These definitions are used in the Vulkan headers, so they are guaranteed
to be available on all platforms that the CTS supports. Trying to define
things in terms of other types is unnecessarily complicated. This is a
first small step towards using the standard types everywhere in the CTS.

Component: Framework
Affects: *

Change-Id: I492dff86e4f6521703feac1d6398d791a579b6fd