Fix for Vulkan 1.3 implementations that don't expose VK_EXT_ycbcr_2plane_444_formats
The VK_EXT_ycbcr_2plane_444_formats extension was only partially promoted
to Vulkan 1.3. In this case the features structure was not promoted
and is not required for Vulkan 1.3 implementations to support it.

This CL modifies the generated framework so that the
VkPhysicalDeviceYcbcr2Plane444FormatsFeaturesEXT is not required to
be supported.

Affects:

dEQP-VK.api.info.get_physical_device_properties2.features
dEQP-VK.*2plane_444*

Components: Vulkan

VK-GL-CTS issue: 3252

Change-Id: I22d95869febd5f8a8c9900e8a659f1609f4c2264