Use the correct attachment parameter when not the default FBO
When using glGetFramebufferAttachmentParameteriv to query the
currently bound framebuffer, the attachment parameter is
different depending on whether the framebuffer is the default
framebuffer or not.

For the default framebuffer GL_DEPTH and GL_STENCIL can be
used, but for non-default these are invalid and
GL_DEPTH_ATTACHMENT and GL_STENCIL_ATTACHMENT must be used.

This CL updates the changes in CL 8180 and CL 8629 to
use the correct attachment parameter when not using
the default framebuffer.

VK-GL-CTS issue: 3092

Affects:
KHR-*.internalformat.renderbuffer.*

Test both with and without: --deqp-surface-type=fbo

Change-Id: Id5f27405861ec62b5fdd84e1a4c2668d8cb272a1