Expand the list of format for image store tests
None of these formats have a SPIR-V format qualifier, so they are only
supported in formatless tests, but feature_flags2 allows these formats
to be supported on some implementations.

The structure of the tests means that it's mostly the store tests that
are expanded -- most of the other types require format qualifiers for
most of the testing.

Components: Vulkan
Affects: dEQP-VK.image.store.without_format.*
         dEQP-VK.image.format_reinterpret.*

Change-Id: I4f2f728b730177d4aa4cda4cdca4feb0d5ee7ad7