Increase tolerance in multiple interpolation tests
These tests test that if a shader has multiple varyings with different
interpolation modes each varying is interpolated using the correct
interpolation mode.

The test runs a shader with multiple varyings with different
interpolation modes and expects this to be an exact match with an image
produced by a similar shader with a single varying of the same
interpolation mode.

Due to floating point inaccuracies and different rounding, allow a 1
bit tolerance when comparing the images.

Affects: dEQP-VK.draw.*multiple_interpolation.*

Components: vulkan

VK-GL-CTS issue: 3522

Change-Id: Ib4a8ab527644de5e1c36ecd3e627afe5384cc6f6