Fix arguments for image test reference generation
The default arguments were incorrect in fd819aa, causing some versions
of the generateReferenceImage function to fail.

There were overloads with 2 and 3 arguments, one was added so 3 or 4
arguments were valid, but only the 3 argument version allowed a default.
Thus the 2 argument versions carried on working, but the 3 argument
calls now matched the old 2 argument version with the new parameter,
which mangled the type into a bool and was generally broken. Here the
default is added to the other prototype as well, causing that to be the
preferred overload and everything goes back to how it was before.

Components: Vulkan
VK-GL-CTS issue: 3372
Affects: dEQP-VK.image.format_reinterpret.*
         dEQP-VK.image.* (possibly)

Change-Id: I1d6da992570988018ebff407697c8ef9a62c823f