isExtensionSupported renaming
The purpose of renaming function name to isExtensionStructSupport
is to make it a clearer emphasis on the fact that extension structure
of passed  extension name was promoted in the core and is not guaranteed
to be supported by the device because of the optional nature of some extensions.
In most cases, the function has been replaced with more appropriate
isDeviceFunctionalitySupported.

Components: Vulkan

VK-GL-CTS Issue: 3149

Affect tests: *dEQP-VK.*

Change-Id: Ic0922bee2f1c178e5aaee9065a1d91260477410a