Reduce framebuffer size on some graphicsfuzz tests
Many graphicsfuzz tests are using a 256x256 framebuffer even though all
pixels or almost all pixels are generating the same results. This CL
reduces the framebuffer size to 16x16 for a some of the tests that are
causing problems in some of our test builds.

Components: Vulkan
Affects: dEQP-VK.graphicsfuzz.cov-dfdx-dfdy-after-nested-loops
         dEQP-VK.graphicsfuzz.cov-nested-loop-large-array-index-using-vector-components
         dEQP-VK.graphicsfuzz.cov-nested-loops-different-iteration-rates-function-copy-array-elements
         dEQP-VK.graphicsfuzz.cov-nested-loops-set-struct-data-verify-in-function

Change-Id: I25fbbb09473d86140a23b1218db344484b563653