Fix memory leaks when creating test hierarchy
The lack of a virtual destructor in one of the classes used as part of
some test parameters made createDescriptorValveMutableTests() leak
memory whenever it was called, in the moment the test hierarchy is being
destroyed. The leak will happen not only when running mutable descriptor
test, but when the binding_model test group needs to be generated for
any reason, including running tests in adjacent groups.

Hence, the change has no specific list of affected tests. At the same
time, no test result should be affected by the change.

Components: Vulkan
VK-GL-CTS issue: 3583

Change-Id: I7e847b5927359cef27f9992cd53991dd1a7cb781