Apply BLAS pool to ray_tracing_pipeline.build.* tests
In addition to replacing traditional memory allocation
with a pool of BLAS structures the tests have been
explicitly divided into GPU and CPU parts. The main
motivation to do this was a inability to run the tests
on implementations that do not support both methods
concurently. so the names have changed respectively.
Nevertheless,the logic of the tests remained unchanged.

Affects: dEQP-VK.ray_tracing_pipeline.build.*

Components: Vulkan

VK-GL-CTS issue: 3599

Change-Id: I7fd022797b1f0fbea9c59bab7ca580d5d0a3eb62