Use getDevice() instead of m_logicalDevice.get() so it works in both paths
This CL fixes a regression introduced with CL9062. For the non-dynamic
paths of the test the m_logicalDevice does not contain a valid
device handle. The newly added getDevice() already handles both
the dynamic and non-dynamic paths, so use that instead of
m_logicalDevice.get().

Affects:

dEQP-VK.memory.binding.*

Components: Vulkan

VK-GL-CTS Issue: 3627

Change-Id: I4ebf74a000553d70db86035ac50b2d4f3257d6ac