Fix ShaderBallotFunctionRead test
The current version of the test draws at 2x2 quad, expecting the value of the
4 pixels to be the same because readFirstInvocationARB is used.

The problem is that depending on the resolution used: if the 4 fragments are
processed in more than one sub-group, then the test will fail.
For instance, on AMD hw the test would pass using the default resolution, but
fail for 256x256.

Fix this by drawing in the top-left corner instead of the middle of the screen.

VK-GL-CTS public issue: 251
Components: OpenGL

Affects:
KHR-GL*shader_ballot_tests.ShaderBallotFunctionRead

Change-Id: Ib182976aeeaab1380d3a5e1cad97fe898cbcdb8e