Remove unused code from integer function tests
The shaderType was being passed to various functions because of the old
GL idea that precision might mean different things in different stages.
This has never been true in Vulkan, so that code is not required.

Also, remove code for printing floats and bools, which is not used in
the integer function tests.

Components: Vulkan
Affects: dEQP-VK.glsl.builtin.function.integer.*

Change-Id: I45cec040eb9ac2bd45f5b7a1845a4937ff7e9576