Split pipeline.txt mustpass into multiple files
Scripts that generate mustpass files were changed to allow
specification of test groups that should be spitted to separate
.txt files. With this change python dictionaries are heavily
used (not just lists) and because of that order of groups in
all .txt files has changed. (Dictionaries are sorted)
This change uses new functionality for pipeline group that now
will be splited to 3 files (each representing subgroup).

Components: Framework

VK-GL-CTS issue: 3641

Affects:
*

Change-Id: I8e339fa71cb2348a5074f53697f9d283c9ba9c35