Remove unneeded check for graphicsPipelineLibraryFastLinking
Graphics pipeline tests using libraries that are linked into final
pipelines without using CREATE_LINK_TIME_OPTIMIZATION_BIT_EXT were
checking if the graphicsPipelineLibraryFastLinking property was set to
true. This is not needed, because

(a) it's a property and not a feature and
(b) linking pipelines without link-time optimization is always allowed.

This commit removes the check and allows more graphics pipeline library
tests run on a wider variety of implementations.

In addition, checkPipelineLibraryRequirements was checking if the
implementation supported VK_KHR_pipeline_library, but this type of
dependencies are checked in dEQP-VK.info.device_extensions.

Affected tests:
dEQP-VK.*fast_linked_library*

VK-GL-CTS issue: 3743
Components: Vulkan

Change-Id: I4c67c195b87a4ab1c24d5d28caf54ea565551b14