Corrected shaders to draw uniform colors.
For MSAA, the subsample of the pixel may be in the triangle, but the
center of the pixel is outside the triangle. This pixel depth may be out
of the depth range and gl_FragCoord.z is negative, especially when the
triangle is narrow and the depth gradient is large. This will draw an
unexpected black for the pixel after the calculation.

This only happens for configs that are not in the must-pass:
'--deqp-surface-width=3840 --deqp-surface-height=2160
--deqp-gl-config-name=rgba8888d24s8ms4'

Components: OpenGL ES
VK-GL-CTS issue: 4110
Affects: dEQP-GLES3.functional.occlusion_query.*

Change-Id: Ib92fe4b90e326bc8a313f123b1df1cdc4417f6d1