Adds missing multisample state structs.
When compiling a fragment shader pipeline library, the fragment shader
state is defined by, among other state, the
VkPipelineMultisampleStateCreateInfo if sample shading is enabled.

In the affected tests, a fragment shader module was being compiled
which accessed builtins decorated with SampleId. This in turn enabled
sample shading, and in such cases, the
VkPipelineMultisampleStateCreateInfo is potentially required by
drivers for their internal logic. For this reason the spec mandates
the struct's inclusion in the fragment shader state when sample
shading is enabled.

Unfortunately it's not simple to change the GraphicsPipelineWrapper to
avoid this erroneous configuration, since in general it would require
parsing the SPIR-V to check if the fragment shader is accessing any
input builtins with the problematic decorations.

Instead a validation run was done over
dEQP-VK.pipeline.pipeline_library.* to ensure there aren't other
callsites falling into this trap, and no other cases were found.

Components: Vulkan

Affects:
dEQP-VK.pipeline.pipeline_library.multisample.*

VK-GL-CTS issue: 3975

Change-Id: I33a6b391af800eb6c23d846b70a4aac2387a74e3