Fix BufferView access tests for scaled types
The way in which these tests were written assumes the input and output
to the shaders are of the same type; however, for scaled types they are
converted from integral to floating point before being processed in the
shader. This change corrects that issue as well as adds the two missing
scaled types to the framework.

Affected tests:
dEQP-VK.api.buffer_view.access.uniform_texel_buffer.*

VK-GL-CTS Issue: 4239

Components: Vulkan

Change-Id: I84f904648c6f2cf301111250dba28feabf782788