Test EDS3 feature interactions
EDS3 does not have any mandatory features but, if supported, some of
them make other features or extensions mandatory. The
device_mandatory_features test has been modified to reflect this.

As part of these changes, the framework generation script now
understands a way to express an extension is mandatory if a feature is
present, which was not possible before.

Affects:
dEQP-VK.info.device_mandatory_features

Components: Vulkan
VK-GL-CTS issue: 4073

Change-Id: I4647881fe5d124666bf0bc3241b2884a8171de33