Treat VulkanSC faults as test failures
Add a fault callback, that records all faults passed through it. This
is added to the default device and custom devices.
After each test, check the recorded faults and also call vkGetFaultData.
Log all faults to the qpa file, and set the test to fail if there are
any non-WARNING faults.

Fix the out of date callback type in vkDefs.hpp, so no casts are required.

Affects: dEQP-VKSC.*
VK-GL-CTS issue: 4291

Change-Id: Iffe6fadf158eef1e90ba5012a564a4234f1555ee