Modify the max_vertex tests to have room for all attributes
Currently they use the maximum number of vec4 attributes, plus
VertexIndex and InstanceIndex. This doesn't fit in the pipeline memory
when the number of attributes is the largest possible. The spec says
that builtins should be counted towards these limits as well, so reduce
the number of attributes slightly to fit properly.

Components: Vulkan
VK-GL-CTS issue: 4379
Affects: dEQP-VK.pipeline.*.vertex_input.max_attributes.*

Change-Id: I482d5862b1ef303557e13411d4610fe3cc47026b