Fix drmFd for negative display tests
These tests try to pass an invalid drm FD and expect to see specific
error codes and return values as a result. For this purpose, the tests
pass "FD + 1" for the drmFD parameter, where FD is a valid FD for the
physical device. This, however, will likely generate a valid FD for the
same device on implementations that call to dup() on the physical device
FD to create an internal copy of the file descriptor. Fix this by
creating an invalid FD from a valid file system path instead.

Components: Vulkan
VK-GL-CTS issue: 4377
Affects: dEQP-VK.wsi.acquire_drm_display.*_fd

Change-Id: I5fc958903c9415039249b03681da399352714171