Fix generation of vkEntryPointValidation.inl
This change fixes a mistake that was done during rebase of CL10558.
Probably during resolving conflicts line that executes writeInlFile
in gen_framework.py was grabbed from writeEntryPointValidation
function to be used in writeGetDeviceProcAddr function.
As a result vkEntryPointValidation.inl was not saved. This was
undiscovered as this file is not changing and it is saved in
generated folder on the repository. But in cases where build saves
.inl files to different locations lack of vkEntryPointValidation.inl
will result in compilation error.

Components: Framework

VK-GL-CTS public issue: 396

Affects:
dEQP-VK.*

Change-Id: I63a5f830606e81a4528a4df4478bacab6a160dde