Test clearing images with 64-bit formats
When dealing with 64-bit formats, one of the problems we have is that
the clear color values in VkClearColor are 32-bit long. The spec has
been recently modified to properly specify what should happen in those
cases: implementations are supposed to reinterpret the first 2 array
elements as a single 64-bit value for the Red component, and the same
for the last 2 elements and the Green component. Clearing of 64-bit
formats with additional components is not supported.

New tests:
dEQP-VK.api.image_clearing.*r64*

Affects:
dEQP-VK.api.image_clearing.*

Components: Vulkan
VK-GL-CTS issue: 3369

Change-Id: I8d35a6f4ddec39fbb1a47e165690d5f790cf7e7a