Fix generation of knownConformanceVersions
Specification CI does a sparse clone of CTS and
because of that it is missing "external\vulkancts\doc" folder
that is needed to generate vkKnownConformanceVersions.inl.
This change detects this in gen_framework.py and prints
a warning instead of trying to continue generation of
vkKnownConformanceVersions.inl file. Note that generation of
this file is not using any data from vulkan-docs.

Components: Framework

VK-GL-CTS issue: 4511

Affects:
dEQP-VK.api.driver_properties.conformance_version

Change-Id: I2ac2983f97b7a0cb3874936bb2704658dcbc2b48