surfaceless: Fix shared contexts and implement makeCurrent
Previously, the surfaceless platform ignored the shared context passed
in to EglRenderContext, instead always passing in EGL_NO_CONTEXT as the
shared context argument to eglCreateContext. However, some tests attempt
to share data between contexts, which is invalid usage of the EGL API.
To fix, the surfaceless platform needs to respect the shared context
argument, passing in the appropriate context to eglCreateContext as
required in EGL.

This fixes GL_INVALID_VALUE errors from glDeleteProgram in some tests
when run with Mesa drivers, as they may delete programs in a different
context than they are created.

In addition, makeCurrent is not implemented for surfaceless, causing two
new tests that use it to fail when run under the surfaceless platform,
spitting the following error in the logs:

  RenderContext::makeCurrent() is not implemented at gluRenderContext.cpp:49

This commit also implements the function, resolving the test failure
under surfaceless.

Affects:
KHR-GLES31.core.compute_shader.simple-compute-shared_context*

Components: Framework
VK-GL-CTS issue: 4515

Change-Id: I1482ef952e7f03a2738a0353a4f4708a935797ed