Add bottom-edge check to FSR attachment rate tests
These tests iterate over a triangle, checking that the number of
fragments sharing the same atomic value matches the fragment rate.
The test skips fragments on the edge of the triangle as they could be
partially covered.

The tests are currently missing a check for the bottom edge, which
is causing failures when testing the 1x4 fragment rate.

Add an additional check to skip any fragments that lie on the bottom
edge of the triangle.

Affects:
dEQP-VK.fragment_shading_rate.renderpass2.attachment_rate.*
dEQP-VK.fragment_shading_rate.dynamic_rendering.attachment_rate.*

Components: Vulkan

VK-GL-CTS issue: 4537

Change-Id: I85fa578daec134842108e65a620dc5b98d7fb74c