Check acceleration structure creation does not touch buffers
This commit adds some new tests that verify the acceleration structure
creation function (vkCreateAccelerationStructureKHR) does not touch the
creation buffer, and the buffer contents are only modified when the
acceleration structure is built.

For this, the tests reuse a buffer in the creation of a top-level or
bottom-level acceleration structure and check the structure which was
created and built first is still usable and drawing with it still works.

New tests:
dEQP-VK.ray_tracing_pipeline.misc.reuse_creation_buffer*

VK-GL-CTS issue: 4548
Components: Vulkan

Change-Id: Ieffee346df765be214981cc433e368f59500f211