Fix image_to_image tests to actually check copying
Source and destination images are filled with the same
data, and therefore, even if no copy happens the test
will pass due to that. This change fill dst with white
or red that is different from the gradient used for the
source.

Affects: dEQP-VK.api.copy_and_blit.*.image_to_image.*

Components: Vulkan
VK-GL-CTS issue: 4487

Change-Id: Ibdb7939b8877b16ca7bf0975d8666106cc06b427