Fix incorrect query result of array size
The issue is that a resource is declared as an array with length 5
and actually only the first element of the array is used, cts expects
the active resource’s array length be 5. Some optimizations are done
and un-used program resources are eliminated at shader compile time in our compiler so array length of 1 is reported by our driver (because only the first element of the array is used and our compiler treats only one element is active for this array). This is valid per spec.

Actually for this case, based on spec, reporting any value between 1 and 5 is valid. To make the test behavior consistent, we use a dynamic index to refer the array, thus all compiler should treat all elements in the array are active, then 5 should be the only expected value.

Affected tests:
KHR-GLES31.core.program_interface_query.uniform-types
KHR-GLES31.core.program_interface_query.arrays-of-arrays

Components: OpenGL ES

VK-GL-CTS issue: 4594

Change-Id: Ibf4556709a777b6f3233d1835afdd4d61e35a569