Fix EGL config selection in wide_color tests
These tests use eglChooseConfig but expect the returned config to
precisely match the requested channel sizes, which is incorrect.

The issue was noticed on a system that exposes both RGB10A2 and RGBA10
configs. This breaks the 1010102 tests, which will incorrectly select
RGBA10 instead of RGB10A2 and then fail.

Using eglu::chooseConfig instead of eglChooseConfig fixes this issue
and allows some redundant code to be removed from the tests.

Affected tests:
dEQP-EGL.functional.wide_color.*

Change-Id: I3e40cbbcd90044e81f3b12b22385be717d6ec1cf