Fix depth bias constant factor test
The affected test uses a D16_UNORM depth buffer and overwrites its
contents with an initial draw of a full screen quad in which the vertex
depth has values from 0.3 to 0.5. Once the depth buffer has been
"cleared" in this way, it uses a second pipeline to draw the same quad
again, but this time it applies a depth bias, with a depth bias constant
factor of 1.0. In theory, this should slightly increase the depth value
of each rasterization point compared to the previous draw, and the depth
test is configured such that the depth comparison op. is "greater than".
If everything goes right, the second draw should completely cover the
first one. Both draws use different geometry colors, so the expected
result is that the whole framebuffer contains the second color.

However, in practice the test can fail. A depth bias constant factor of
1.0 means that, at each sampling point, the depth value is increased by
the constant called "r" in the spec. Without the depth bias control
extension, the value of "r" depends on the depth attachment format.
Quoting the spec:

> ...
> Otherwise its value is implementation-dependent but must be at most
>    r = 2 × 2^(-n)
> where n is the number of bits used for the depth aspect when using a
> fixed-point attachment

It turns out that RADV is using r=2^(-n) for these cases. However,
increasing the depth value by adding r, with a depth bias constant
factor of 1.0, may not result in an increase of the depth buffer value,
which makes the "greater than" test fail. Take the following example:

In a D16_UNORM attachment, value 0 represents 0.0 and value 65535
represents 1.0. Each unit increase in the depth attachment value
represents an increase of 1.5259021896696422e-05 in depth.

Value 0.5 cannot be represented exactly in this case. 32767 would
represent 0.49999237048905165 and 32768 would represent
0.5000076295109483. Both values are within 7.629510948348184e-06 of 0.5.
Lets suppose the implementation rounds up, or rounds to even, and
chooses 32768. Then, in the second draw 0.5 is increased by r=2-16,
resulting in a depth value of 0.5000152587890625. Which depth attachment
value represents most faithfully the increased depth value? 32768, with
an error of 2.2888300010848184e-05. 32769 would mean an error of
3.814732190754455e-05 (higher).

This shows that both the old depth value and the increased depth value
could be represented by 32768, failing the "greater than" test. To avoid
this issue, the depth bias constant factor is increased to 2.0.

Affects:
dEQP-VK.rasterization.depth_bias.d16_unorm_constant_one_greater

Components: Vulkan
VK-GL-CTS issue: 4694

Change-Id: I901162e0144eebc9098fb30d9336e4eb0e8187c8