Do not use and check for STORAGE image support, when it is not used in the test
This reverts 37e64193 and provides alternate fix, without making the
tests not supported on some platforms.

Storing to the image is not used in this test, so it does not need to
be checked for support and requested for images.

VK-GL-CTS Issue: 4667

Affected tests:
dEQP-VK.sparse_resources.image_block_shapes.*

Components: Vulkan
Change-Id: I0e11b15042fc5687482a1ed103c1acae16f6d40a